Refactor controller into vlc_player.dart
