+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: flutter_vlc_player
 description: A new Flutter project.
-version: 0.1.2
+version: 0.1.3
 author: Solid Software <hello@solid.software>
 homepage: https://github.com/solid-software/flutter_vlc_player
